+++ tests/TestCase/Command/PhpDocMakerCommandTest.php
@@ -28,6 +28,11 @@ class PhpDocMakerCommandTest extends TestCase
     /**
      * @var string
      */
+    protected $source = TEST_APP;
+
+    /**
+     * @var string
+     */
     protected $target = TMP . 'output' . DS;
 
     /**
@@ -38,8 +43,7 @@ class PhpDocMakerCommandTest extends TestCase
     {
         parent::tearDown();
 
-        unlink_recursive($this->target);
-        @unlink(TESTS . DS . 'test_app' . DS . 'php-doc-maker.xml');
+        @unlink(TEST_APP . 'php-doc-maker.xml');
     }
 
     /**
@@ -48,12 +52,9 @@ class PhpDocMakerCommandTest extends TestCase
      */
     public function testExecuteOptions()
     {
-        $source = TESTS . DS . 'test_app';
-
         $Command = new PhpDocMakerCommand();
-        $Command->PhpDocMaker = $this->getPhpDocMakerMock();
         $Command->PhpDocMaker = $this->getMockBuilder(PhpDocMaker::class)
-            ->setConstructorArgs(compact('source'))
+            ->setConstructorArgs([$this->source, $this->target])
             ->setMethods(['build'])
             ->getMock();
         $commandTester = new CommandTester($Command);
@@ -65,7 +66,7 @@ class PhpDocMakerCommandTest extends TestCase
             'target' => $this->target,
             'title' => null,
         ];
-        $commandTester->execute(compact('source') + ['--target' => $this->target]);
+        $commandTester->execute(['source' => $this->source, '--target' => $this->target]);
         $this->assertFalse($commandTester->getOutput()->isVerbose());
         $this->assertEquals($expectedOptions, $commandTester->getInput()->getOptions());
 
@@ -76,7 +77,8 @@ class PhpDocMakerCommandTest extends TestCase
             'target' => $this->target,
             'title' => 'A project title',
         ];
-        $commandTester->execute(compact('source') + [
+        $commandTester->execute([
+            'source' => $this->source,
             '--debug' => true,
             '--target' => $this->target,
             '--title' => 'A project title',
@@ -91,13 +93,13 @@ class PhpDocMakerCommandTest extends TestCase
             'no-cache' => false,
             'target' => $this->target,
         ];
-        create_file($source . DS . 'php-doc-maker.xml', '<?xml version="1.0" encoding="UTF-8" ?>
+        create_file($this->source . DS . 'php-doc-maker.xml', '<?xml version="1.0" encoding="UTF-8" ?>
 <php-doc-maker>
     <title>My test app</title>
     <target>' . $this->target . '</target>
     <verbose>true</verbose>
 </php-doc-maker>');
-        $commandTester->execute(compact('source'));
+        $commandTester->execute(['source' => $this->source]);
         $this->assertTrue($commandTester->getOutput()->isVerbose());
         $this->assertEquals($expectedOptions, $commandTester->getInput()->getOptions());
 
@@ -106,13 +108,13 @@ class PhpDocMakerCommandTest extends TestCase
             'debug' => true,
             'no-cache' => true,
         ] + $expectedOptions;
-        create_file($source . DS . 'php-doc-maker.xml', '<?xml version="1.0" encoding="UTF-8" ?>
+        create_file($this->source . DS . 'php-doc-maker.xml', '<?xml version="1.0" encoding="UTF-8" ?>
 <php-doc-maker>
     <title>My test app</title>
     <target>' . $this->target . '</target>
     <debug>true</debug>
 </php-doc-maker>');
-        $commandTester->execute(compact('source'));
+        $commandTester->execute(['source' => $this->source]);
         $this->assertTrue($commandTester->getOutput()->isVerbose());
         $this->assertEquals($expectedOptions, $commandTester->getInput()->getOptions());
     }
@@ -123,12 +125,10 @@ class PhpDocMakerCommandTest extends TestCase
      */
     public function testExecute()
     {
-        $source = TESTS . DS . 'test_app';
-
         $Command = new PhpDocMakerCommand();
         $Command->PhpDocMaker = $this->getPhpDocMakerMock();
         $commandTester = new CommandTester($Command);
-        $commandTester->execute(compact('source') + ['--target' => $this->target]);
+        $commandTester->execute(['source' => $this->source, '--target' => $this->target]);
 
         //Tests output
         $output = $commandTester->getDisplay();
@@ -137,7 +137,7 @@ class PhpDocMakerCommandTest extends TestCase
         $this->assertRegExp('/Founded \d+ functions/', $output);
         $this->assertRegExp('/Elapsed time\: \d+\.\d+ seconds/', $output);
 
-        $this->assertStringContainsString('Sources directory: ' . $source, $output);
+        $this->assertStringContainsString('Sources directory: ' . $this->source, $output);
         $this->assertStringContainsString('Target directory: ' . $this->target, $output);
         $this->assertStringContainsString('Rendered index page', $output);
         $this->assertStringContainsString('Rendering functions page', $output);
@@ -152,8 +152,7 @@ class PhpDocMakerCommandTest extends TestCase
      */
     public function testExecuteMissingComposerAutoloader()
     {
-        $Command = new PhpDocMakerCommand();
-        $commandTester = new CommandTester($Command);
+        $commandTester = new CommandTester(new PhpDocMakerCommand());
         $commandTester->execute(['--debug' => true, 'source' => TMP]);
         $this->assertSame(1, $commandTester->getStatusCode());
         $output = $commandTester->getDisplay();
@@ -166,10 +165,9 @@ class PhpDocMakerCommandTest extends TestCase
      */
     public function testExecuteOnError()
     {
-        $source = TESTS . DS . 'test_app';
         $Command = new PhpDocMakerCommand();
         $Command->PhpDocMaker = $this->getMockBuilder(PhpDocMaker::class)
-            ->setConstructorArgs(compact('source'))
+            ->setConstructorArgs([$this->source, $this->target])
             ->setMethods(['build'])
             ->getMock();
 
@@ -187,19 +185,19 @@ class PhpDocMakerCommandTest extends TestCase
         }));
 
         $commandTester = new CommandTester($Command);
-        $commandTester->execute(['--debug' => true] + compact('source'));
+        $commandTester->execute(['source' => $this->source, '--debug' => true]);
         $this->assertSame(1, $commandTester->getStatusCode());
         $this->assertStringContainsString('[ERROR] A notice error...', $commandTester->getDisplay());
         $this->assertStringContainsString(sprintf('On file `%s`', __FILE__), $commandTester->getDisplay());
 
-        $commandTester->execute(['--debug' => true] + compact('source'));
+        $commandTester->execute(['source' => $this->source, '--debug' => true]);
         $this->assertSame(1, $commandTester->getStatusCode());
         $output = $commandTester->getDisplay();
         $this->assertStringContainsString('[ERROR] Something went wrong...', $output);
         $this->assertStringContainsString(sprintf('On file `%s`', $expectedException->getFile()), $output);
         $this->assertStringContainsString(sprintf('line %s', $expectedException->getLine()), $output);
 
-        $commandTester->execute(['--debug' => true] + compact('source'));
+        $commandTester->execute(['source' => $this->source, '--debug' => true]);
         $this->assertSame(0, $commandTester->getStatusCode());
         $this->assertStringContainsString('[OK] Done!', $commandTester->getDisplay());
     }